[Swift in WebKit] Add a Swift module to bmalloc to better support Swift on Cocoa platforms https://bugs.webkit.org/show_bug.cgi?id=301054 rdar://162959248
Reviewed by Elliott Williams and Mike Wyrzykowski.

Enable Swift and Swift-Cxx interop in bmalloc. In addition to enabling the possibility of using
Swift directly in bmalloc, it serves two important immediate-term purposes:

1. The module verifier tool does not currently support libraries, such as WTF and bmalloc. Compiling a
Swift file in bmalloc itself allows for at least a minimal means of verification (specifically, successful
compilation of this file indicates a necessary but not sufficient condition).

2. Enabling interop in bmalloc is an effective stepping stone to more complex projects like WTF
and WebCore, particularly because it is lower on the WebKit stack.

* Source/bmalloc/Configurations/bmalloc.xcconfig:
* Source/bmalloc/Configurations/module.modulemap:
* Source/bmalloc/bmalloc.swift: Added.
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/302570@main