[iOS] Remove use of `-[UIScreen mainScreen]` in `WKImagePreviewViewController` https://bugs.webkit.org/show_bug.cgi?id=302017 rdar://164101198
Reviewed by Abrar Rahman Protyasha.

`-[UIScreen mainScreen]` was deprecated in iOS 26. Work towards removing its
use in WebKit.

* Source/WebKit/UIProcess/WKImagePreviewViewController.h:
* Source/WebKit/UIProcess/WKImagePreviewViewController.mm:
(-[WKImagePreviewViewController initWithCGImage:defaultActions:elementInfo:presentingView:]):

Pass the presenting view to the initializer so that a `UIScreen` can be obtained
from the view's window. Since this view controller is presented as a result of a
context menu interaction, the view must be in the view hierarchy.

The view controller's own view is not used, since it is not in the hierarchy
during initialization. Additionally, `-viewIsAppearing` and `-viewDidAppear`
is too late to set the preferred content size.

(-[WKImagePreviewViewController initWithCGImage:defaultActions:elementInfo:]): Deleted.
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView assignLegacyDataForContextMenuInteraction]):
(-[WKContentView continueContextMenuInteraction:]):
(-[WKContentView _presentedViewControllerForPreviewItemController:]):

Canonical link: https://commits.webkit.org/302604@main