WebExtensionDeclarativeNetRequestSQLiteStore::addRules returns without calling the completion handler https://bugs.webkit.org/show_bug.cgi?id=302032 rdar://164109895
Reviewed by Timothy Hatcher and Brian Weinstein.

We shouldn't be returning here without calling the completion handler. We wouldn't hit this case
for empty rules since we have checks in place before we reach this point. I think the only way we'd
hit this is if we somehow get a corrupted value back after we parse the JSON. We could return
here and call the completion handler with an error, but I think a continue is sufficient.

* Source/WebKit/UIProcess/Extensions/WebExtensionDeclarativeNetRequestSQLiteStore.cpp:
(WebKit::WebExtensionDeclarativeNetRequestSQLiteStore::addRules):

Canonical link: https://commits.webkit.org/302609@main