Crash at WebKit::WebExtension::populateActionPropertiesIfNeeded https://bugs.webkit.org/show_bug.cgi?id=302028 rdar://164059373
Reviewed by Timothy Hatcher.

We were unconditionally grabbing the value off the icon result, regardless if there was an error or
not. Fix this by only grabbing the result if there is no error.

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtension.mm
* Source/WebKit/UIProcess/Extensions/WebExtension.cpp:
(WebKit::WebExtension::populateActionPropertiesIfNeeded):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtension.mm:
(TestWebKitAPI::TEST(WKWebExtension, ActionParsing)):
Verified this test crashing without the change, and passes with the change.

Canonical link: https://commits.webkit.org/302611@main