[libpas] Fix ExpendableMemoryTests.cpp https://bugs.webkit.org/show_bug.cgi?id=301938 rdar://152525465
Reviewed by Dan Hecht.

We're getting a failure at an assertion for
`bmalloc_get_allocation_size(largeArray) == 2016`.
The reason for this is that we changed the size-class boundary for
small-medium to ~1616B, so the size-class width for a 2016B allocation
goes up from 16B to 128B, and thus the actual bucket should be the 2048
size-class.

Canonical link: https://commits.webkit.org/302615@main