REGRESSION(302506@main): Triggered assertions on visionOS Debug tester https://bugs.webkit.org/show_bug.cgi?id=302013 rdar://164096712
Reviewed by Ryosuke Niwa.

As of 302223@main, subclasses of CanMakeCheckedPtr must be heap-allocated. If they are not (and are
not exempted), CanMakeCheckedPtrBase::~CanMakeCheckedPtrBase crashes due to an assertion failure.
Resolved this by heap-allocating FullScreenWindowControllerVideoPresentationModelClient and storing
it in a unique_ptr in WKFullScreenWindowController.

* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController initWithWebView:]):
(-[WKFullScreenWindowController _isBestVideoInFullScreen]):
(-[WKFullScreenWindowController bestVideoFullscreenModeChanged]):
(-[WKFullScreenWindowController fullScreenViewControllerDidInvalidate:]):
(-[WKFullScreenWindowController fullScreenViewController:bestVideoPresentationInterfaceDidChange:]):

Canonical link: https://commits.webkit.org/302631@main