REGRESSION(302582@main): TestWebKitAPI.WebKit2.CrashGPUProcessAfterApplyingConstraints times out rdar://164194814 https://bugs.webkit.org/show_bug.cgi?id=302113
Reviewed by Jean-Yves Avenard.

Before 302582@main, we were using a single CoreAudioCaptureUnit so getting the deviceID in MockAudioCaptureInternalUnit::set was ok.
Now that we use on macOS a different CoreAudioCaptureUnit when echoCancellation is off, we have the following issue:
- TestWebKitAPI.WebKit2.CrashGPUProcessAfterApplyingConstraints starts capturing microphone.
- Echo cancellation is turned off
- GPUProcess is terminated and will restart to restart capturing microphone
- When restarting to capture, the default CoreAudioCaptureUnit is not set up as echoCancellation is off, so the deviceID is not set
- This triggers a debug assert in MockAudioCaptureInternalUnit::set.

To fix that issue, we make sure in CoreAudioCaptureSource::createForTesting to populate the microphone deviceID, that we set in MockRealtimeMediaSourceCenter for microphones.
This ensures that MockAudioCaptureInternalUnit::set is provided that ID.
The MockAudioCaptureInternalUnit can then get the device using MockRealtimeMediaSourceCenter::mockMicrophoneFromDeviceID.

Canonical link: https://commits.webkit.org/302722@main