[Tools][GTK] Performance tests fail if WebKitTestRunner prints something in stderr https://bugs.webkit.org/show_bug.cgi?id=302152
Reviewed by Fujii Hironori.

Since we migrated to the new SDK the GTK Performance bot has been failing
all the performance tests in the step `perf-test` because there are some
Mesa related warnings printed and the tooling is marking as failed any
test that has stderr output not matching the ignore filters.

So ignore this warnings by adding them to the list of filters, and also
share the list with the WPE port.

* Tools/Scripts/webkitpy/performance_tests/perftest.py:
(PerfTest):

Canonical link: https://commits.webkit.org/302728@main