[JSC] GreedyRegAlloc: adjust the spill cost heuristic https://bugs.webkit.org/show_bug.cgi?id=302159 rdar://164259068
Reviewed by Keith Miller and Yusuke Suzuki.

The spillCostSizeBias parameter had the wrong type, causing it to
always be 1 when the intention was for it to be 1000. Regardless,
both 1 and 1000 perform similarly and are too small.

It's better to decrease the influence of the live range size (which
is the sum of the distances of the intervals that make up the range)
since this is just a crude approximation for the degree of interference.

Also, add a new register allocator metric that helped guide the
selection of this value.

Tests: covered by existing JSC tests
Canonical link: https://commits.webkit.org/302735@main