[ macOS Tahoe ] fast/repaint/block-inputrange-repaint.html is a constant text failure https://bugs.webkit.org/show_bug.cgi?id=301000 rdar://162884649
Reviewed by Wenson Hsieh.

With the redesigned controls in macOS Tahoe, the size of the slider thumb
changed. This means that the repaint rect expected by the test is no longer
accurate.

Fix by using `getBoundingClientRect()` to determine the size and position
of the thumb, which provides an accurate repaint rect.

* LayoutTests/fast/repaint/block-inputrange-repaint-expected.txt:
* LayoutTests/fast/repaint/block-inputrange-repaint.html:

A "tolerance" of 2px is used to inflate the repaint rect. This is necessary
since some ports (macOS) adjust the repaint rect to account for shadows.

* LayoutTests/platform/glib/fast/repaint/block-inputrange-repaint-expected.txt: Removed.
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/302739@main