Always set lockdown_mode_byte, hardened_byte in pas_mte_config.c https://bugs.webkit.org/show_bug.cgi?id=301565 rdar://163557770
Reviewed by Mark Lam.

Right now, there are certain code paths (privileged processes) that do
not ever set the lockdown mode byte or the hardened byte. This does not
have any effect at present because both of these bytes are only ever
used *within* WebContent to up the hardening level, but it’s confusing /
bug prone in the case that someone adds some condition that checks for
those bytes elsewhere in the code.

This patch fixes that by lifting the ldm-check and setting hardened_byte
in the privileged-process codepath.

Canonical link: https://commits.webkit.org/302741@main