[WKCI] `architectures` for queues should be concatenated with `-`, not space. https://bugs.webkit.org/show_bug.cgi?id=301952 rdar://164027693
Reviewed by Jonathan Bedard.

When uploading build products from CI to S3, if there are multiple
architectures specified in the queue information, we upload with a space in the
file name (as the concatenator of the architectures) instead of a dash.

This PR changes this to avoid issues with spaces in file names.

* Tools/CISupport/build-webkit-org/factories.py:
    (Factory.__init__): Change space to `-`.
* Tools/CISupport/ews-build/steps.py:
    (ConfigureBuild.__init__): Ditto.
* Tools/Scripts/bisect-builds:
    (QueueDescriptor):
        -> (__init__): Ditto.
        -> (pretty_string): Ditto, but to `/` instead.
* Websites/webkit.org/wp-content/themes/webkit/build-archives.php: Replace `%20` with `-` in `$platforms`.

Canonical link: https://commits.webkit.org/302748@main