RELEASE_ASSERT() under WebCore::Page::serviceWorkerGlobalObject https://bugs.webkit.org/show_bug.cgi?id=302189 rdar://164059424
Reviewed by Youenn Fablet and Timothy Hatcher.

We're hitting an assertion inside Page::serviceWorkerGlobalObject() because
WebExtensions code is calling this function with an isolated world and this
is something we don't support and have never supported.

To address the issue, I am bailing out at the call site when the world is
not the normal one.

* Source/WebKit/WebProcess/Extensions/WebExtensionContextProxy.cpp:
(WebKit::WebExtensionContextProxy::enumerateFramesAndNamespaceObjects):

Canonical link: https://commits.webkit.org/302758@main