[GTK][WPE] Performance test Animation/balls.html does not work because of XHTML 1.0 Strict doctype headers https://bugs.webkit.org/show_bug.cgi?id=302182
Reviewed by Fujii Hironori.

The strict XHTML doctype header forces the browser to parse the inline
JavaScript as XML unless it is wrapped inside CDATA sections via comments,
which isn't happening in this case.

So, for example, "angle < Math.PI" breaks parsing because < appears to start
an XML tag which is not finished.

Instead of adding the CDATA section comments, directly switch the header to
modern HTML that does not have this issues with the parsing.

* PerformanceTests/Animation/balls.html:

Canonical link: https://commits.webkit.org/302759@main