Web Inspector: REGRESSION(?): cannot stop Timeline recording after submitting a form https://bugs.webkit.org/show_bug.cgi?id=301643
Reviewed by Darin Adler.

In the case of a cross-origin form submission, the old page is destroyed so all of the timeline instruments are gone and won't notify Web Inspector that they're no longer recording.

We effectively should reset the count of active timeline instruments, but it's not even really necessary to have that value in the first place when we already have a `WI.TimelineManager.CapturingState`.

* Source/WebInspectorUI/UserInterface/Controllers/TimelineManager.js:
(WI.TimelineManager):
(WI.TimelineManager.prototype.capturingStarted):
(WI.TimelineManager.prototype.capturingStopped):

Canonical link: https://commits.webkit.org/302776@main