Crash: com.apple.WebKit.WebContent at AVFCore:  -[AVAssetWriterInputHelper requestMediaDataWhenReadyOnQueue:usingBlock:] https://bugs.webkit.org/show_bug.cgi?id=302226 rdar://164271457
Reviewed by Youenn Fablet and Eric Carlson.

Don't attempt to write to the AVAssetWriter when closing when an error occurred.
From the documentation "Before calling this method, you must ensure that the receiver is attached to an AVAssetWriter via a prior call to -addInput: and that -startWriting has been called on the asset writer."
so also add handling to ignore any remaining samples if it's not possible to use the writer any longer.

* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterAVFObjC.h:
* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterAVFObjC.mm:
(WebCore::MediaRecorderPrivateWriterAVFObjC::addAudioTrack):
(WebCore::MediaRecorderPrivateWriterAVFObjC::addVideoTrack):
(WebCore::MediaRecorderPrivateWriterAVFObjC::allTracksAdded):
(WebCore::MediaRecorderPrivateWriterAVFObjC::close):

Canonical link: https://commits.webkit.org/302779@main